unexport interface implementation struct, move components to the right packages
